retry use memo and succeed
use memo to implement left recursive, only a memo in a left recursion is needed.
very simple solution! pass 3 tests.
test A: Ax|a
test A: Bx|a; B:A|b
test A: Bx|a; B:C; C:A|b
I think that it should be enough to process most of cases.